feat: divide prerendering-related plugins into separate files
This will enable developers to pick which plugins they need. Also, this makes it much easier to pick all except one, which needs to be modified by the dev.
