+++ packages/api-page-builder/src/prerendering/hooks/index.ts
@@ -1,0 +1,15 @@
+import afterMenuUpdate from "./afterMenuUpdate";
+import afterPageDelete from "./afterPageDelete";
+import afterPagePublish from "./afterPagePublish";
+import afterPageUnpublish from "./afterPageUnpublish";
+import afterSettingsUpdate from "./afterSettingsUpdate";
+
+export default () => {
+    return [
+        afterMenuUpdate(),
+        afterPageDelete(),
+        afterPagePublish(),
+        afterPageUnpublish(),
+        afterSettingsUpdate()
+    ];
+};
